Fix c_int to u32 type mismatch in c-wrapper-crate FFI wrapper (#38)
* fix: change value() return type from u32 to i32 to match c_int FFI return type


---------

+++ c-wrapper-crate/src/lib.rs
@@ -11,6 +11,6 @@ unsafe fn bar() -> core::ffi::c_int {
     unreachable!()
 }
 
-pub fn value() -> u32 {
-    (unsafe { bar() }) as u32
+pub fn value() -> i32 {
+    unsafe { bar() }
 }
